requirements.txt
fastapi
openai
langchain
uvicorn
python-dotenv
boto3
slack_sdk
jira
streamlit
gitpython
prometheus-api-client
